third_party/blink/web_tests/fast/forms/select/menulist-field-sizing-autofill.html

<!DOCTYPE html>
<link rel="help" href="https://github.com/whatwg/html/pull/9903">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
#target { field-sizing: content; }
</style>
<select id="reference"><option><option>foo<option>bar<option>foobar</select>
<select id="target"><option><option>foo<option>bar<option>foobar</select>

<script>
test(() => {
  assert_less_than(target.offsetWidth, reference.offsetWidth);
  internals.setSuggestedValue(target, 'foobar');
  assert_equals(target.offsetWidth, reference.offsetWidth);
}, 'field-sizing:content is not applied to autofilled controls');
</script>
